CardData.cs - separate initialization into two parts
Some card behaviours reference other cards. To do so, they go to Cards.FromId/etc. For this method to work, data must be constructed. Therefore we can't initialize card behaviours until CardDefs is constructed. We do this entirely in a static constructor so no one else has to do anything, it 'just works'.